Refactorization complete
