--- order-service/src/main/java/com/ecommerce/orderservice/application/dto/UserAuthenticationDetails.java
@@ -1,3 +1,0 @@
-package com.ecommerce.orderservice.application.dto;
-
-public record UserAuthenticationDetails(String token, String userId) {}
