--- order-service/src/main/java/com/ecommerce/orderservice/application/port/in/CancelOrderUseCase.java
@@ -1,7 +1,0 @@
-package com.ecommerce.orderservice.application.port.in;
-
-import java.util.UUID;
-
-public interface CancelOrderUseCase {
-    void cancelOrder(UUID id);
-}
