+++ order-service/src/main/java/com/ecommerce/orderservice/application/service/OrderApplicationService.java
@@ -1,0 +1,10 @@
+package com.ecommerce.orderservice.application.service;
+
+import com.ecommerce.orderservice.domain.port.in.OrderUseCase;
+
+/**
+ * Application service interface for order management use cases.
+ * Implements the OrderUseCase inbound port.
+ */
+public interface OrderApplicationService extends OrderUseCase {
+}
