+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/adapter/feign/ProductFeignAdapter.java
@@ -1,8 +1,8 @@
 package com.ecommerce.orderservice.infrastructure.adapter.feign;
 
 import com.ecommerce.orderservice.application.dto.*;
-import com.ecommerce.orderservice.application.port.out.ProductServicePort;
 import com.ecommerce.orderservice.domain.exception.OrderValidationException;
+import com.ecommerce.orderservice.domain.port.out.ProductServicePort;
 import com.ecommerce.shared.exception.ExceptionError;
 import feign.FeignException;
 import io.github.resilience4j.circuitbreaker.annotation.CircuitBreaker;
