+++ order-service/src/main/java/com/ecommerce/orderservice/infrastructure/adapter/security/dto/UserAuthenticationDetails.java
@@ -1,0 +1,6 @@
+package com.ecommerce.orderservice.infrastructure.adapter.security.dto;
+
+/**
+ * DTO for transferring user authentication details between the security adapter and application layer.
+ */
+public record UserAuthenticationDetails(String token, String userId) {}
